real_estate_agent/requirements.txt
agent-framework-core
agent-framework-foundry
azure-identity
python-dotenv
uvicorn
aiohttp